LayoutTests/platform/ios-sim-deprecated/transforms/3d/hit-testing/hover-rotated-negative-z-expected.txt

CONSOLE MESSAGE: line 15: TypeError: null is not an object (evaluating 'hit.id')
transformed
Test passes if the hover state of a transformed div is activated on the entire surface of the div.

Element at 70, 70 has id "transformed": PASS
